Upgrade deprecated rollup-plugin-babel to @rollup/babel-plugin
This requires to explicitly set the option `babelHelpers` that decides how to handle them in the final bundle. The plugin's readme suggests using 'runtime' for library packages (which this is). This requires us to use the @babel/plugin-transform-runtime plugin and to specify @babel/runtime as a dependency of the package. See https://github.com/rollup/plugins/tree/master/packages/babel#babelhelpers